db: fix SA4011 remove ineffective break statement
It's unclear whether Monitor is supposed to only log errors or log an
error and quit when watchdog test query fails. As used in main.go the
returned error is not checked, so we continue to use it to log
errors.

Monitor's functionality might be redundant with the DB heartbeat
check. In the past aggressively restarting the server instead of
waiting for the DB to become available again hurt availability.