Preserve changes backed out from main
The changes in 7bc7ab2ce78848cd6a69dc4753990a6d89e02eac were backed out
in order to ensure the production branch is deployable from the head.
(These changes had been landed for testing, and deployed for a while in
stage, but never validated. Hence we were tagging the prior revision for
production deploys.)

The other files in the original commit were improvements to the
documentation and development processes. Since these did not impact the
production code, they were left in.