adding contentsignaturepki leaf cert CN overrides
This adds the ability to override the common name
set in the Subject of the leaf certificates
generated by the contentsignaturepki type singer
configurations.

This is to allow for new certificates created with
the same key material but different singer
configurations to be validated by Firefox and
other clients as they were previously. This is an
important requirement for the add-on and content
signing root key succession plan as we're asking
the distribution services to use new signers (new
`keyids`) for the new keys, while keeping the old
common names in the end-user clients.

Fixes [bug 1886886](https://bugzilla.mozilla.org/show_bug.cgi?id=1886886)