make test exemplary
The docs for SubdomainOverride
(subdomain_override) now state that if you add it
to a signer config, you should also change the X5U
and ChainUploadLocation configs to avoid colliding
with the other signer. And then we do that in the
tests even though this is not part of the codebase
that is in a position to check for that kind of
behavior.