use explicit struct tags for YAML decoding
This makes it clear what the actual YAML configs
will contain, clearing up some of the underscore
confusion we had previously. No changes have been
made to what the config file schemas are.

The `SingerOpts` type doesn't seem to have a
MarshalJSON or equivalent for YAML, so I suspect
this field never gets unmarshalled from the
configs at all. But, in an effort to minimize
change, I've left it in.