avoid destructive gnupg clean up
The deletion of `~/.gnupg` in the Makefile seems to
have been mooted by more careful use of gpg's
`--homedir` flag during the history of autograph.

Not being able to run the tests locally without
fear of blowing away your own gpg set up is not
great and slows down progress in working with
autograph.

If we find inconsistently failing tests due to
this change, we should apply more `--homedir` and
equivalent work to autograph and its tests.