gpg2: Remove deprecated --secret-keyring option
The --secret-keyring option has been silently ignored since GPG v2.1,
so it never did anything anyways, and in more recent builds this now
generates a console warning. We should simply remove it.

This should be fine so long as we are making use of the --homedir
option to prevent unexpected changes to the user's GPG keyring.