Fix command substitution bug in lint and static checks
It seems that the make's $(shell ...) function is not evaluated quite
when we expect it, freuqently leading to the check of the results before
we even run the linter. To avoid this problem, we should try escaping the
dollar-sign so that the underlying shell does the grep command
substitution instead of make.

This fixes the bug where you see a spurious warning about
/tmp/autograph-golint.txt not existing and a failure to catch linter
errors in CI.