Fix failures in CI testing (#894)
* Remove golang.org/x/tools/cmd/cover tooling

This was deprecated in Go 1.5, and I think it now exists as a builtin
command.

* Install goveralls v0.0.11 for coverage testing.

We fix the use of version v0.0.11 as that seems to match up with the
packages in go.mod, and we switch to 'go install' from 'go get' since
it's a testing tool rather than a build dependency.

* Unit tests should now pass on Circle CI

* Install golint version v0.0.0-20190409202823 for linting.

We use version v0.0.0-20190409202823-959b441ac422 as this seems to match
up with the packages in go.mod and go.sum, and we switch go 'go install'
from 'go get' since it's a testing tool rather than a build dependency.

* Install staticcheck version v0.2.2 for testing

We use version v0.2.2 as this seems to be the latest version which will
build cleanly on Go 1.16. We also switch to 'go install' from 'go get'
since it's a testing tool rather than a build dendency.

* Fix command substitution bug in lint and static checks

It seems that the make's $(shell ...) function is not evaluated quite
when we expect it, freuqently leading to the check of the results before
we even run the linter. To avoid this problem, we should try escaping the
dollar-sign so that the underlying shell does the grep command
substitution instead of make.

This fixes the bug where you see a spurious warning about
/tmp/autograph-golint.txt not existing and a failure to catch linter
errors in CI.

* Fix linter warnings

* golint, vet, and gofmt checks should pass now

* Remove conflicting vendored copy of contentsignature verifier

The static checker is finding errors about the verifier/contentsignature
module, because it is checking against a stale version in the vendor
directory rather than the one included in the sources. We should fix
this so there is only one source of truth for the code being compiled.

* Fix autograph-monitor unit test breakage

This used to pass only because of the broken vendoring of the
verifier/contentsignature package, and the tests were written to match
the stale vendored version.

* Static check tests should pass now

* All tests should pass now