remove nested go.mod for verifier/contentsignature (#899)
This was causing surprises for devs making changes
in that package and hitting a variety of build
problems when they realized they hadn't updated
the top-level go.mod. And other hacks failed to be
kept up (see
https://github.com/mozilla-services/autograph/pull/896#issuecomment-2091770686 ).

The only external user of
verifier/contentsignature was the
mozilla-vpn-client and @oskirby (a maintainer of
it) tells us they've been using a rust version for
a while now. Clients also could just use the
autograph repo's versioning instead.

Can be confirmed with this search:
https://github.com/search?q=autograph%2Fverifier%2Fcontentsignature&type=code