Implement non-PKI contentsignature.VerifyResponse() method
The monitor tests have been failing for quite some time because the
non-PKI contentsignature format isn't required to provide an X5U, but
the verification method assumes there will be one. What we really need
here is a standalone verifier for this signature type that doesn't
depend on, or expect an X5U to be provided.