finish off Go 1.22 upgrade
This patch cleans up some stragglers in the CircleCI config that were
still running on 1.19.

We also specify the patch version of Go in go.mod as that's what the
`1.22` tooling (and, I believe, 1.21) does. If we don't do that, various
third party tools like GitHub's CodeQL can break (see
https://github.com/github/codeql/issues/15647 for the last time I ran
into this). This patch change is actually what clued me into the
CircleCI version bumps that were missing.