finish off Go 1.22 upgrade
This patch cleans up some stragglers in the CircleCI config that were
still running on 1.19.

We also specify the patch version of Go in go.mod as that's what the
`1.22` tooling (and, I believe, 1.21) does. If we don't do that, various
third party tools like GitHub's CodeQL can break (see
https://github.com/github/codeql/issues/15647 for the last time I ran
into this). This patch change is actually what clued me into the
CircleCI version bumps that were missing.

We also had to install go 1.22 into the separate linting job in
CircleCI.

Using the CircleCI go1.22 image there instead of `bookworm` broke
because it couldn't `apt-get update` successfully (something about
`partial` not being right). We then had to use the `BASH_ENV` hack in
there. First, because the `golang-1.22` dpkg doesn't add its `go` to the
PATH, so we need to adjust it ourselves. And then that specific BASH_ENV
hack, because CircleCI's `environment` directives doesn't know what
value `PATH` is, so attempting something like
`${PATH}:/usr/lib/go-1.22/bin` gets you a `PATH` with only the Go 1.22
binaries.

I've made AUT-158 to consolidate our go linting stuff, but some (not
all) of the rationale will be mooted by a move to GitHub Actions.