record statsd metrics for availability (#933)
This records the HTTP response status codes returned by autograph to our
existing statsd infrastructure. It will be helpful for monitoring uptime
and other errors.

Metrics will look like "<api name>.response.status.<status code>" (the
HTTP status code autograph is trying to return) and and "<api
name>.request.attempts" (the number of requests that have come into
autograph from its POV). We also include some
`api.response.status.<status code>` and `api.request.attempts` metrics
to more easily calculate the percentage of successful requests across
all of autograph's public API.

We do so by writing a small middleware that wraps the given
http.ResponseWriter, writes the status code to statsd when `Write` and
`WriteHeader` are called, handles the interaction between `Write` and
`WriteHeader`, and avoids writing the headers and the status code twice.

We also add a mock implemenation of `statsd.ClientInterface` for
the rare cases we want to test what metrics are being sent (which we
have in this patch). It was generated by `mockgen` (sibling to the
uber/mock library added). How to reproduce that mock is included in its
comment at the top of the file (per usual).

Updates AUT-150