move tools/autograph-monitor into top-level make (#943)
This is review feedback from a previous patch.
https://github.com/mozilla-services/autograph/pull/941#discussion_r1705636141

This patches moves tools/autograph-monitor is `PACKAGE_NAMES` in the
top-level Makefile and removes the separated make calls.

Our linting tools (surprisingly!) caught some things in
contentsignaturepki_test.go in autograph-monitor with this change, so we
fix this up, as well. (I suspect this "typedNilNotifier" thing isn't
testing what the authors thought it was, but I'm going for a minimal
change here. I could be told to rip it out in a follow-up.)

This might mess up the coverage calculations we were previously doing,
but I'm not sure we've been getting much value out of those. We can fix
them up later if we do.

This also tries to clean up some unusual ways the autograph-monitor is
referenced in the rest of the Makefile. I'm not sure what that was
about. Perhaps, at some point, the original authors had code files in
that directory with different `package`s in them?