AUT-186: upgrade to aws-sdk-go-v2 (#945)
This turns out to be fairly straightforward? I reworked `uploadToS3` a tiny bit to accommodate basic testing. The easiest way to do this was to move some of the boilerplate one function up. It would be ideal to keep _all_ of it where it currently is, but doing so means we can't dependency inject anything.

Seeing as there's literally no logic other than whether or not errors are forwarded, I'm tempted to drop the tests and move the code back. I could be convinced either way.