consolidate linting and add it to GitHub Actions
This consolidates our linting tools into the golangci-lint tool. It also
removes the unsupported, archived golint tool.

Docs about golangci-lint can be found here: https://golangci-lint.run/

Specifics about how to configure specific linters and what they are can
be found at https://golangci-lint.run/usage/linters/

The new GitHub Action integration also will annotate the PR its run on
with the violations it fails instead of just printing them out to the
console. This means we'll get nice lil UI things directly next to the
code in a PR that's causing the linter violation. This is the Action
maintained by the golangci-lint maintainers.

The default configuration of `golangci-lint` includes `errcheck` which
is very handy but was going to cause a significant amount of churn here.
I've made AUT-202 for us to turn it on later.

We use `depguard` to replace the `check-no-crypto11-in-signers` target.
It's a lil funky (I don't like the "**/" prefix that's required and
could cause false positives if we ever gain another `signer` package
name) but it's nice to have it in place for future checks like it. Plus,
it won't have the previous kind of false positives of the old grep.

`staticcheck` is still run in golangci-lint by default and we make sure
it remains enabled. No change there.

Along the way, pick up some new lint errors that the previous linting
missed.

One is that the `main.httpError` func is Sprintf-like, and `go vet`'s
`printf` checks were being violated in a few places. Same deal with some
uses of `fmt.Errorf` without any additional arguments. Some `log.Printf`
calls had the wrong number of arguments for the format string they were
given. One of the `main.httpError` calls is false positive-y and gets a
useless `"%s"` because we have no "zero arguments" function equivalent,
but it's fine.

Our immediate update of the `debug` variable in `main_racing_test.go`
meant the new `ineffassign` linter was unhappy. It's slightly false
positive-y because I suspect it was folks just trying to be clear about
what the zero value of a bool is, but the new code for that is shorter
and more typical Go.

`go vet` ("`govet`" in golangci-lint parlance) also had an "unreachable
code" trigger in the autograph-client codebase. A nice simplification
made it go away.

`golangci-lint` got a 5 minute timeout (`--timeout`) in the GitHub
Actions. That time was chosen arbitrarily, and 5 minutes would be a long
time to wait for a linting check to finish. We could definitely shorten
it and maybe should.

Also, the Makefile gains the ability to install `golangci-lint`, because
I don't have a better pattern for how folks should install it right now.

Updates AUT-158
Updates AUT-197