move various tools to avoid main func collisions
These Go files all contained `func main`s that conflicted with other
`main`s in the same directory. Only one of these tools are used in our
build or referenced in our docs, and that use (in our softhsm
Dockerfile) is adjusted.

This was done because various tools complain and break. For instance,
`go build ./...`. Now the `...` alias in Go tooling works again.

This was figured out while working on porting the linting work to GitHub
Actions.

Updates AUT-192
Updates AUT-197