Remove tags from deployment job fetch
I think this is a vestigial config from an earlier attempt to roll our
own tags, and they are no longer necessary now that we are using the
docker/metadata-action step. Furthermore, we seem to be failing to
checkout on a github tag event, which is preventing us from shipping
release artifacts.