add integration matrix result job for PR checks
It's annoying to have each individual "Run Integration Tests" matrix job
in the branch protection settings. If we ever need to add or remove one
of those, we'd have to grab a GitHub admin user (`autograph-admins`) to
adjust the protection settings (or, if SRE is allowed to infra-as-code
GitHub, write a separate PR).

Instead, we can add this new integration-test-results job that `needs`
the results of `Run Integration Tests` and, after this merges, we can
add it to the branch protection settings instead of the individual
integration test jobs.

See
https://github.com/orgs/community/discussions/26822#discussioncomment-8285141
and
https://github.com/orgs/community/discussions/26822#discussioncomment-3305794

Updates AUT-151