trim leading slash before uploading chains to S3
The stage environment went down for a while. The details of why can be
found in AUT-231.

One of the triggers was that aws-go-sdk-v2 no longer removes leading
slashes from uploaded S3 key names like the v1 version of the library
did. That caused some cascading failures leading to too many AWS
CloudHSM slots being consumed.

This change trims the leading slashes from the uploaded S3 key names for
the contentsignaturepki certificate chains.

This change also gets us closer to not needing to be careful about
remembering to end chainUploadLocation strings with a slash with an
application of `path.Join`. The `file://` handling elsewhere is still
bugged and so are the places where we do `X5U + chainName`.

Updates AUT-233
Updates AUT-231