remove verifier replace statement in go.mod
Missed when removing the verfifier/contentsignature go.mod in #899.

This has caused other folks to try to add replace statements that they
didn't need to, and it's useless in any case.

Removed the `replace` and ran `go mod vendor` to clean up the vendoring
(plus a `go mod tidy` just in case).