AUT-252: fork crypto11 into the autograph repository (#984)
* AUT-252: fork crypto11 into the autograph repository

This is part 1 of the work for AUT-252, and simply moves `crypto11` from the `vendor` directory into a replacement fork in the root of the Autograph repository. No actual code changes have been made as part of this, but I did:
* Remove `Gopkg.lock` and `Gopkg.yaml`, which are simply unnecessary and possibly confusing.
* Rename the package to `github.com/mozilla-services/autograph/crypto11` to avoid `go mod vendor` copying it into the `vendor` directory, which is annoying and confusing. (Maybe this is the wrong thing to do, in which case I'm happy to revert it and let `go mod vendor` do its thing.)

* fix: address linting issues

With actual fixes where it's easy, and ignoring them where it's not trivial.