upgrade sops to v3.9.1 from v3.4.0
This upgrades sops from v3.4.0 to v3.9.1.

A variety of corrections and improvements have happened since the sops
version we've been shipping on. The most obvious change is the new Go
modules package path.

A number of deprecated, archived, and unused transitive dependencies
were also removed by this upgrade. Seeing autograph stuck on those
unmaintained dependencies (like the v1 aws libraries) through the sops
dependency was the inspiration for this. New transitive deps were added,
per usual.

This patch was created by replacing the sops imports with the new ones
in main.go, and then running:

```
$ go mod tidy
$ go mod vendor
```