upgrade sops to v3
A variety of correction and improvements have happened since the sops
version we've been shipping on.

A number of deprecated, archived, and unused dependencies were also
removed by this upgrade and were the inspiration for this. New
transitive deps were added, per usual.

This patch was created by replacing the sops imports with the new ones
in main.go, and then running:

```
$ go mod tidy
$ go mod vendor
```