correct CheckHSMConnection error handling (#1000)
With GetPrivKeyHandle removed in #992, we forgot to update the error
handling in CheckHSMConnection causing it to always return an error. Now
we rely soely on the earlier PrivateKeyHasPEMPrefix call for whether the
key is in the HSM.

Along the way, we cause main to error if the HSM connection check fails.
This allows autograph and the integration tests to fail immediately if
the HSM is believed to be unavailable.

We found this problem at staging deploy time because we weren't checking
the HSM connection on boot. The deploy was failing because the HSM
healthcheck failed, but those aren't checked in integration tests.