clean up main.initHSM error handling
After #1000, initHSM now has a mix of returning errors and exiting out.
This patch reconciles them into error returns.

We could also do this HSM check for every signer config, but that's
probably better done separately as we're already dealing with boot time
timeout problems in AWS.

Along the way, we note a strange way we need to handle the list of
signer configurations.