monitor: be more lenient and more strict about AUTOGRAPH_URL
Before this patch, AUTOGRAPH_URL was required to include a scheme and a
trailing slash in order to be considered valid but wasn't required to be
an actual URL. This patch defaults the scheme to HTTPS if its not set,
and handles missing trailing slashes.

We use url.JoinPath so that if we forget the trailing slash in the
`AUTOGRAPH_URL` env var for the autograph-monitor, things still work
correctly.

Before this, the autograph-monitor was trying to resolve
`"${AUTOGRAPH_URL}__monitor__"` as a domain name when you forgot the
trailing slash.