bundle top-level aws-sdk-go-v2 in dependabot group (#1022)
Obvious in retrospect, but the top-level github.com/aws/aws-sdk-go-v2
module is not included in the pattern "github.com/aws/aws-sdk-go-v2/*"
for the dependabot group. But it would be nice to bundle it in there. It
usually will show up in that dependabot group PR anyway, but as we've
seen today, it'll also get a separate PR.