remove make-hsm-ee tool
This is tangentially related to AUT-264, where I'm looking at updating any tools that work with HSMs to support GCP. However, it appears that this one is simply unused since everything migrated to `contentsignaturepki`. I can't actually find proof of this, but I can't find any documentation or information on when we would actually use this tool...so, let's remove it?