refresh makecsr and add it to the Dockerfile
The makecsr tool is used to generate a CSR for our addon and
contentsignaturepki intermediates

This patch does a mild refactoring to bring it under test and adds it to
the docker image so it can be used from our private hsm tooling repo.

We also add setting the Subject Alternative Name DNS names to it. This
will make it easier for future tooling that's dealing with Name
Constraints verify it correctly.

Along the way, we also require passing in a signature algorithm.
Signature algorithm, like the name constraints, is something that can be
overridden at CA signing time, but this removes some of the per-cloud
implicit configuration that was in the tool.

Updates AUT-279
Updates AUT-280