AUT-270: implement new deployment workflow for autograph
In AUT-326 we decided that we would:
* Deploy stage in response to new version tags in the autograph repository
* Avoid rebuilding docker images when pushing a version tag to dockerhub (we will instead pull the docker image we deployed to dev from the ref that we tagged)

This commit implements this.

Slightly tangential, but also of note is https://github.com/mozilla-it/global-platform-admin/pull/2138 adding support for disabling automatic sync in ArgoCD, and doing so for autograph prod. This means that prod will notice new version tags, but not automatically deploy in response to them (human approval will be needed).

You can see example runs of this for the push and release cases in:
* https://github.com/bhearsum/autograph/actions/runs/11557423769/job/32167407424
* https://github.com/bhearsum/autograph/actions/runs/11557348483/job/32167157862