monitor: make monitor testable (#1040)
This makes the Handler and monitor funcs accept some arguments so we can
write tests for them. We write one simple golden path test to
demonstrate.

This is an outcome of trying to make the AUTOGRAPH_URL env var handling
better in #1012. That PR will be rebased on to this one.