add benchmarks for GKE random generators
These are to be run in the GKE to confirm my hypothesis about the
performance of the GCP KMS GenerateRandom RPC vs the local random data
generator. I suspect that we're not getting much benefit from the KMS
GenerateRandom (which is called by libkmsp11). The security properties
of it are also not sufficiently different than the modern, very good
random number generators available on Linux in the Go standard library.

Updates AUT-335