count signer requests with some useful tags
This patch records what signers are used, and by whom. This is useful
for a variety of reasons including seeing patterns of problems in
roughly real time and tracking usage of signers.

It also includes whether the user used their "default" signer id
(`keyid` in the API itself). We're trying to move folks away from using
the default signer id and requiring `keyid` to be set. We want that
because we want our clients to be more easily moved from one signer to
another as our keys and algorithm requirements change. (Folks had to
change their code during the add-ons and content-signature root
certificate succession instead of just changing some configs because
they relied on the default signer id.)

Updates AUT-206