monitor: add ability to ignore expiration of signer ids
This patch adds an env var of comma-separated signer ids to the monitor.
The signer ids in the env var will have any expiration date violations
ignored. In the monitor, the expiration violations include the
expiration being close to the current date.

We do this to clear out the monitor errors returned for the
pre-cert-succession signer ids for the remote-settings service. The
remote-settings service won't flip over until mid-January, but we'd like
to have clear monitors before then.

I went back and forth on using an env var or a command-line flag. I
could go either way.

Updates AUT-345