add prometheus and counter for testing
Our statsd integration with prometheus has been turning counters that
should be 1 into 0. It's unclear why this is happening and this patch is
an attempt to debug the issue further.

The foobar_test counter added here is to help determine why. It's
incremented once on boot in both the prometheus and in statsd. We'll
determine if the problem is in prometheus, grafana, telegraf, or in our
statsd client (unlikely this last one) by emitting in both and seeing
how they're captured.

The rest of the org is standardizing on prometheus for metrics up and
down the stack, so we should do the same, anyway. We don't commit to
migrating all of the existing metrics here, but make it plausible to do
so.

This adds the port 2112 to Dockerfile for autograph as the place where
the prometheus metrics are served from.

Follow up patches in our infra-as-code will add the fetching of
prometheus metrics.

Updates AUT-393
Updates AUT-199