migrate to github.com/DataDog/datadog-go/v5/statsd (#1080)
This migrates us to the v5 major version channel of DataDog's statsd
client. Specifically, we bump to v5.6.0.

This was done by running:

```
$ gofmt -w -r '"github.com/DataDog/datadog-go/statsd" -> "github.com/DataDog/datadog-go/v5/statsd"' .
$ go mod tidy
$ go mod vendor
```

Then, using the new client consturction api.

We set the namespace and don't set the buflen size anymore, as the
default is now optimized for each transport. See
statsd.WithBufferPoolSize for more information on that.

Along the way, we updated our mockstatsd with mockgen. See the command
in that comments of that generated file to reproduce that work.