migrate to go.uber.org/gomock
The github.com/golang/mock project is no longer maintained and has been archived.

The go.uber.org/mock project is a fork of golang/mock that is actively
maintained.

So, we do the work to recreate the mocks with the new library. The
internal/mockstatsd one was already using the go.uber.org/mock library
and so doesn't need to change.